Add Conversational Bot (#51)
* add data generation notebook

* refactor

* Add itent detection support. Implement using Azure Conversations API.

* add chat service and test

* move argument parsing

* (wip) implementation for dub, translate. refactor transcribe.

* add voice to dub, fix integration issues

* rewrite chatbot test; address review comments

* remove test data notebook; will turn into code in a separate PR

+++ freespeech/lib/gdocs.py
@@ -21,7 +21,7 @@ class Page:
     voice: Character
     clip_id: str
     method: Source
-    original_audio_level: float
+    original_audio_level: int
     video: str | None
 
 
@@ -150,7 +150,7 @@ def parse_properties(text: str) -> Page:
         if attribute not in keys and attribute != "video":
             raise TypeError(f"{attribute} must be defined")
 
-    properties["original_audio_level"] = float(properties["original_audio_level"])
+    properties["original_audio_level"] = int(properties["original_audio_level"])
     properties["video"] = None if "video" not in keys else properties["video"] or None
 
     return Page(**properties)
